+++ src/lib.rs
@@ -7,10 +7,10 @@
 //! - [Handlers](#handlers)
 //!     - [Debugging handler type errors](#debugging-handler-type-errors)
 //! - [Routing](#routing)
-//!     - [Precedence](#precedence)
 //!     - [Matching multiple methods](#matching-multiple-methods)
 //!     - [Routing to any `Service`](#routing-to-any-service)
 //!         - [Routing to fallible services](#routing-to-fallible-services)
+//!     - [Wildcard routes](#wildcard-routes)
 //!     - [Nesting routes](#nesting-routes)
 //! - [Extractors](#extractors)
 //!     - [Common extractors](#common-extractors)
@@ -177,75 +177,8 @@
 //!
 //! You can also define routes separately and merge them with [`Router::or`].
 //!
-//! ## Precedence
-//!
-//! Note that routes are matched _bottom to top_ so routes that should have
-//! higher precedence should be added _after_ routes with lower precedence:
-//!
-//! ```rust
-//! use axum::{
-//!     body::{Body, BoxBody},
-//!     handler::get,
-//!     http::Request,
-//!     Router,
-//! };
-//! use tower::{Service, ServiceExt};
-//! use http::{Method, Response, StatusCode};
-//! use std::convert::Infallible;
-//!
-//! # #[tokio::main]
-//! # async fn main() {
-//! // `/foo` also matches `/:key` so adding the routes in this order means `/foo`
-//! // will be inaccessible.
-//! let mut app = Router::new()
-//!     .route("/foo", get(|| async { "/foo called" }))
-//!     .route("/:key", get(|| async { "/:key called" }));
-//!
-//! // Even though we use `/foo` as the request URI, `/:key` takes precedence
-//! // since its defined last.
-//! let (status, body) = call_service(&mut app, Method::GET, "/foo").await;
-//! assert_eq!(status, StatusCode::OK);
-//! assert_eq!(body, "/:key called");
-//!
-//! // We have to add `/foo` after `/:key` since routes are matched bottom to
-//! // top.
-//! let mut new_app = Router::new()
-//!     .route("/:key", get(|| async { "/:key called" }))
-//!     .route("/foo", get(|| async { "/foo called" }));
-//!
-//! // Now it works
-//! let (status, body) = call_service(&mut new_app, Method::GET, "/foo").await;
-//! assert_eq!(status, StatusCode::OK);
-//! assert_eq!(body, "/foo called");
-//!
-//! // And the other route works as well
-//! let (status, body) = call_service(&mut new_app, Method::GET, "/bar").await;
-//! assert_eq!(status, StatusCode::OK);
-//! assert_eq!(body, "/:key called");
-//!
-//! // Little helper function to make calling a service easier. Just for
-//! // demonstration purposes.
-//! async fn call_service<S>(
-//!     svc: &mut S,
-//!     method: Method,
-//!     uri: &str,
-//! ) -> (StatusCode, String)
-//! where
-//!     S: Service<Request<Body>, Response = Response<BoxBody>, Error = Infallible>
-//! {
-//!     let req = Request::builder().method(method).uri(uri).body(Body::empty()).unwrap();
-//!     let res = svc.ready().await.unwrap().call(req).await.unwrap();
-//!
-//!     let status = res.status();
-//!
-//!     let body = res.into_body();
-//!     let body = hyper::body::to_bytes(body).await.unwrap();
-//!     let body = String::from_utf8(body.to_vec()).unwrap();
-//!
-//!     (status, body)
-//! }
-//! # }
-//! ```
+//! Routes are not allowed to overlap and will panic if an overlapping route is
+//! added. This also means the order in which routes are added doesn't matter.
 //!
 //! ## Routing to any [`Service`]
 //!
@@ -376,6 +309,41 @@
 //! See ["Error handling"](#error-handling) for more details on [`handle_error`]
 //! and error handling in general.
 //!
+//! ## Wildcard routes
+//!
+//! axum also supports wildcard routes:
+//!
+//! ```rust,no_run
+//! use axum::{
+//!     handler::get,
+//!     Router,
+//! };
+//!
+//! let app = Router::new()
+//!     // this matches any request that starts with `/api`
+//!     .route("/api/*rest", get(|| async { /* ... */ }));
+//! # async {
+//! # axum::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
+//! # };
+//! ```
+//!
+//! The matched path can be extracted via [`extract::Path`]:
+//!
+//! ```rust,no_run
+//! use axum::{
+//!     handler::get,
+//!     extract::Path,
+//!     Router,
+//! };
+//!
+//! let app = Router::new().route("/api/*rest", get(|Path(rest): Path<String>| async {
+//!     // `rest` will be everything after `/api`
+//! }));
+//! # async {
+//! # axum::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
+//! # };
+//! ```
+//!
 //! ## Nesting routes
 //!
 //! Routes can be nested by calling [`Router::nest`](routing::Router::nest):
@@ -409,6 +377,25 @@
 //! have the matched prefix stripped. This is necessary for services like static
 //! file serving to work. Use [`OriginalUri`] if you need the original request
 //! URI.
+//!
+//! Nested routes are similar to wild card routes. The difference is that
+//! wildcard routes still see the whole URI whereas nested routes will have
+//! the prefix stripped.
+//!
+//! ```rust
+//! use axum::{handler::get, http::Uri, Router};
+//!
+//! let app = Router::new()
+//!     .route("/foo/*rest", get(|uri: Uri| async {
+//!         // `uri` will contain `/foo`
+//!     }))
+//!     .nest("/bar", get(|uri: Uri| async {
+//!         // `uri` will _not_ contain `/bar`
+//!     }));
+//! # async {
+//! # axum::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
+//! # };
+//! ```
 //!
 //! # Extractors
 //!